Integration of different trakt api(WIP)

+++ app/src/main/kotlin/com/ashish/movieguide/ui/common/rating/RatingChangeObserver.kt
@@ -8,11 +8,11 @@ import javax.inject.Singleton
 @Singleton
 class RatingChangeObserver @Inject constructor() {
 
-    private val ratingChangeSubject = PublishSubject.create<Pair<Long, Double>>()
+    private val ratingChangeSubject = PublishSubject.create<Pair<Long, Int>>()
 
-    fun getRatingObservable(): Observable<Pair<Long, Double>> = ratingChangeSubject.hide()
+    fun getRatingObservable(): Observable<Pair<Long, Int>> = ratingChangeSubject.hide()
 
-    fun notifyRatingChanged(mediaId: Long, rating: Double) {
+    fun notifyRatingChanged(mediaId: Long, rating: Int) {
         ratingChangeSubject.onNext(Pair(mediaId, rating))
     }
 }
