+++ app/src/main/kotlin/com/ashish/movieguide/ui/common/rating/RatingMvpView.kt
@@ -4,5 +4,5 @@ import com.ashish.movieguide.ui.base.mvp.MvpView
 
 interface RatingMvpView : MvpView {
 
-    fun showSavedRating(rating: Double?)
+    fun showSavedRating(rating: Int?)
 }
